+++ 05-SceneManager/PlayScene.h
@@ -40,7 +40,7 @@ protected:
 
 public:
 	bool isTurnOnCamY = false;
-	float multiScene = 1.65;
+	double  multiScene = 1.65;
 	CPlayScene(int id, LPCWSTR filePath);
 
 	virtual void Load();
